+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+### Added
+- added support for retry attempts using `--retry-attempts`
+
 ## 2.2.1 - 2023-07-17
 ### Fixed
 - Local section links are no longer rendered as broken relative links (e.g. `[this section](#section-header)`)